\name{testipakettiArma-package}
\alias{testipakettiArma-package}
\alias{testipakettiArma}
\docType{package}
\title{
\packageTitle{testipakettiArma}
}
\description{
\packageDescription{testipakettiArma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testipakettiArma}
\packageIndices{testipakettiArma}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{testipakettiArma}

Maintainer: \packageMaintainer{testipakettiArma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
